chore(refactor) updating the code

+++ server/controllers/propertyController.js
@@ -3,18 +3,8 @@ import Authenticator from '../helper/authenticator';
 import pool from '../config/connection';
 import AuthValidator from '../middlewares/authValidator';
 
-const { decodeToken, generateToken } = Authenticator;
+const { decodeToken } = Authenticator;
 const { header } = AuthValidator;
-// const header = (req) => {
-//   const token = req.headers.authorization.split(' ')[1] || req.headers.authorization || req.headers['x-access-token'] || req.headers.token || req.body.token;
-//   const decoded = decodeToken(token);
-//   return decoded;
-// };
-// const header = (req) => {
-//   const token = req.headers.authorization || req.headers['x-access-token'] || req.headers.token || req.body.token;
-//   const decoded = decodeToken(token);
-//   return decoded;
-// };
 let property;
 /**
  * Defines methods for properties
@@ -40,14 +30,13 @@ class PropertyController {
       const {
         status, price, state, city, address, type, image_url, property_name, image_url_2, image_url_3, description, map_lat, map_lng, purpose,
       } = req.body;
-      const sqlQuery = `INSERT INTO properties(owner, property_name, status, type, state, city, address, price, image_url, image_url_2, image_url_3, owner_email, owner_phone, purpose, description, map_lat, map_lng)
-                    VALUES($1, $2, $3, $4, $5, $6, $7, $8, $9, $10, $11, $12, $13, $14, $15, $16, $17)
+      const sqlQuery = `INSERT INTO properties(owner, property_name, type, state, city, address, price, image_url, image_url_2, image_url_3, owner_email, owner_phone, purpose, description, map_lat, map_lng)
+                    VALUES($1, $2, $3, $4, $5, $6, $7, $8, $9, $10, $11, $12, $13, $14, $15, $16)
                     RETURNING *`;
-      const values = [owner, property_name, status, type, state, city, address, price, image_url, image_url_2, image_url_3, owner_email, owner_phone, purpose, description, map_lat, map_lng];
+      const values = [owner, property_name, type, state, city, address, price, image_url, image_url_2, image_url_3, owner_email, owner_phone, purpose, description, map_lat, map_lng];
       property = await client.query({ text: sqlQuery, values });
       if (property.rows && property.rowCount) {
         property = property.rows[0];
-        // const token = await generateToken(122);
         // return res.status(201).json({ status: 'success', data: { id: property.id, status: property.status, type: property.type, state: property.state, city: property.city, address: property.address, price: property.price, image_url: property.image_url } });
         return res.status(201).json({ status: 'success', data: property });
       }
@@ -68,8 +57,6 @@ class PropertyController {
      * @memberof PropertyController
      */
   static async getProperties(req, res) {
-    // const { token } = header(req);
-  // const decoded = decodeToken(token);
     const { type } = req.query;
     const sqlQuery = 'SELECT * FROM properties ORDER BY created_on ASC';
     const sqlQueryType = 'SELECT * FROM properties WHERE type = $1 ORDER BY created_on DESC';
@@ -100,10 +87,6 @@ class PropertyController {
      * @memberof PropertyController
      */
   static async getProperty(req, res) {
-    // const owner = header(req).id; const owner_phone = header(req).phone_number;
-    // const owner_email = header(req).email;
-    // const { token } = header(req);
-    // const token = req.headers.authorization || req.headers['x-access-token'] || req.headers.token || req.body.token;
     const { propertyId } = req.params;
     const client = await pool.connect();
     try {
@@ -112,7 +95,6 @@ class PropertyController {
       property = await client.query({ text: sqlQuery, values });
       if (property.rowCount) {
         property = property.rows[0];
-        // const token = await generateToken(122);
         return res.status(200).json({ status: 'success', data: property });
       }
       return res.status(404).json({ status: 'error', error: 'Property Not Found' });
@@ -133,15 +115,15 @@ class PropertyController {
      */
   static async updateProperty(req, res) {
     const { propertyId } = req.params;
-    // const ownerId = header(req).id;
-    // const token = req.headers.authorization || req.headers['x-access-token'] || req.headers.token || req.body.token;
-    const { token } = header(req);
+    const token = header(req);
+    const decoded = decodeToken(token);
+    const ownerId = decoded.payload.id;
     const client = await pool.connect();
-    const findOneQuery = 'SELECT * from properties WHERE id = $1';
+    const findOneQuery = 'SELECT * from properties WHERE id = $1 AND owner = $2';
     const sqlQuery = `UPDATE properties SET property_name = $1, status = $2, type = $3, state = $4, city = $5, address = $6, price = $7, image_url = $8, image_url_2 = $9, image_url_3 = $10, purpose = $11, description = $12, map_lat = $13, map_lng = $14
-                      WHERE id = $15 RETURNING *`;
+                      WHERE id = $15 AND owner = $16 RETURNING *`;
     try {
-      property = await client.query(findOneQuery, [propertyId]);
+      property = await client.query(findOneQuery, [propertyId, ownerId]);
       if (!property.rows[0]) {
         return res.status(404).json({ status: 'error', error: 'Property Not Found' });
       }
@@ -161,6 +143,7 @@ class PropertyController {
         req.body.map_lat || property.rows[0].map_lat,
         req.body.map_lng || property.rows[0].map_lng,
         propertyId,
+        ownerId,
       ];
       property = await client.query(sqlQuery, values);
       property = property.rows[0];
@@ -181,21 +164,18 @@ class PropertyController {
      * @memberof PropertyController
      */
   static async deleteProperty(req, res) {
-    // const owner = header(req).id; const owner_phone = header(req).phone_number;
-    // const owner_email = header(req).email;
-    // const token = req.headers.authorization || req.headers['x-access-token'] || req.headers.token || req.body.token;
-    const { token } = header(req);
+    const token = header(req);
+    const decoded = decodeToken(token);
+    const ownerId = decoded.payload.id;
     const { propertyId } = req.params;
-    // const ownerId = header(req).id;
-    const deleteQuery = 'DELETE FROM properties WHERE id = $1 RETURNING *';
+    const deleteQuery = 'DELETE FROM properties WHERE id = $1 AND owner = $2 RETURNING *';
     const client = await pool.connect();
     try {
-      property = await client.query(deleteQuery, [propertyId]);
+      property = await client.query(deleteQuery, [propertyId, ownerId]);
       if (!property.rows[0]) {
         return res.status(404).json({ status: 'error', error: 'Property Not Found' });
       }
-      // const token = await generateToken(122);
-      return res.status(200).json({ status: 'success', data: { token, message: `Property with ID: ${propertyId} deleted` } });
+      return res.status(200).json({ status: 'success', data: { message: `Property with ID: ${propertyId} deleted` } });
     } catch (err) {
       return res.status(500).json({ status: 'error', error: 'Internal Server Error' });
     } finally {
@@ -212,57 +192,22 @@ class PropertyController {
      * @memberof PropertyController
      */
   static async markProperty(req, res) {
-    // const owner_phone = header(req).phone_number;
-    // const owner_email = header(req).email;
-    // const token = req.headers.authorization || req.headers['x-access-token'] || req.headers.token || req.body.token;
-    const { token } = header(req);
     const { propertyId } = req.params;
-    // const ownerId = header(req).id;
-    const findOneQuery = 'SELECT * from properties WHERE id = $1';
-    const sqlQuery = 'UPDATE properties SET status = $1 WHERE id = $2 RETURNING *';
+    const token = header(req);
+    const decoded = decodeToken(token);
+    const ownerId = decoded.payload.id;
+    const findOneQuery = 'SELECT * from properties WHERE id = $1 AND owner = $2';
+    const sqlQuery = 'UPDATE properties SET status = $1 WHERE id = $2 AND owner = $3 RETURNING *';
     const client = await pool.connect();
     try {
-      property = await client.query(findOneQuery, [propertyId]);
+      property = await client.query(findOneQuery, [propertyId, ownerId]);
       if (!property.rows[0]) {
         return res.status(404).json({ status: 'error', error: 'Property Not Found' });
       }
-      const values = [req.body.status || property.rows[0].status, propertyId];
+      const values = [req.body.status || property.rows[0].status, propertyId, ownerId];
       property = await client.query(sqlQuery, values);
       const propert = property.rows[0];
-      // const token = await generateToken(122);
-      return res.status(200).json({ status: 'success', data: { token, created_on: propert.created_on } });
-    } catch (err) {
-      return res.status(500).json({ status: 'error', error: 'Internal Server Error' });
-    } finally {
-      await client.release();
-    }
-  }
-
-  /**
-     * Edit property price
-     * @static
-     * @param {object} req - request
-     * @param {object} res - response
-     * @returns
-     * @memberof PropertyController
-     */
-  static async editPropertyPrice(req, res) {
-    // const owner_phone = header(req).phone_number;
-    // const owner_email = header(req).email;
-    const { propertyId } = req.params;
-    // const ownerId = header(req).id;
-    const findOneQuery = 'SELECT * from properties WHERE id = $1';
-    const sqlQuery = 'UPDATE properties SET price = $1 WHERE id = $2 RETURNING *';
-    const client = await pool.connect();
-    try {
-      property = await client.query(findOneQuery, [propertyId]);
-      if (!property.rows[0]) {
-        return res.status(404).json({ status: 'error', error: 'Property Not Found' });
-      }
-      const values = [req.body.price || property.rows[0].price, propertyId];
-      property = await client.query(sqlQuery, values);
-      const token = await generateToken(122);
-      return res.status(200).json({ status: 'success', data: [token, property.rows[0]] });
+      return res.status(200).json({ status: 'success', data: propert });
     } catch (err) {
       return res.status(500).json({ status: 'error', error: 'Internal Server Error' });
     } finally {
@@ -279,11 +224,13 @@ class PropertyController {
      * @memberof PropertyController
      */
   static async getPropertiesByAgent(req, res) {
-    // const ownerId = header(req).id;
+    const token = header(req);
+    const decoded = decodeToken(token);
+    const ownerId = decoded.payload.id;
     const sqlQuery = 'SELECT * FROM properties WHERE owner = $1 ORDER BY created_on ASC';
     const client = await pool.connect();
     try {
-      property = await client.query(sqlQuery, [1]);
+      property = await client.query(sqlQuery, [ownerId]);
       if (property.rowCount) {
         return res.status(200).json({ status: 'success', data: property.rows });
       }
